requirements/base.txt
django==1.11
django-bootstrap-form==3.4
django-contrib-comments==1.8.0
django-notifications-hq==1.3
django-rest-auth==0.9.3
django-allauth==0.35.0
django-ipware==2.0.1
djangorestframework-jwt==1.11.0
django-mptt==0.9.0
pycodestyle==2.3.1
djangorestframework==3.8.0
pillow==5.0.0
django-filter==1.1.0
markdown==2.6.11
coreapi==2.3.3
factory-boy==2.10.0
djang-model-utils==3.1.1
django-cors-headers==2.2.0

git+https://github.com/zmrenwu/django-activity-stream.git@master#egg=django-activity-stream
